Refactor: abstract parsing and handling commands to CommandParser class
