+++ src/main/java/CancelGPT.java
@@ -8,6 +8,7 @@ public class CancelGPT {
     private final String CHATBOT_NAME;
     private final TasksList TASKS_LIST;
     private TasksStorage tasksStorage;
+    private CommandParser commandParser;
 
     public CancelGPT() {
         this.CHATBOT_NAME = "CancelGPT";
@@ -18,6 +19,7 @@ public class CancelGPT {
             System.out.println("Unable to use TASKS STORAGE. Exiting program");
             System.exit(1);
         }
+        this.commandParser = new CommandParser(this);
     }
 
     public String getName() {
@@ -69,157 +71,11 @@ public class CancelGPT {
     }
 
     public void handleCommand(String command) {
-        try {
-            if (command.equals(Command.LIST.toString())) {
-                displayTasksList();
-            } else if (command.startsWith(Command.DELETE.toString())) {
-                int taskNumber = parseDeleteTaskCommand(command);
-                deleteTask(taskNumber);
-            } else if (command.startsWith(Command.MARK.toString())) {
-                int taskNumber = parseMarkTaskCommand(command);
-                markTask(taskNumber);
-            } else if (command.startsWith(Command.UNMARK.toString())) {
-                int taskNumber = parseUnmarkTaskCommand(command);
-                unmarkTask(taskNumber);
-            } else if (command.startsWith(Command.TODO.toString())) {
-                Task toDoTask = parseToDoTaskCreationCommand(command);
-                handleAddingTask(toDoTask);
-            } else if (command.startsWith(Command.DEADLINE.toString())) {
-                Task deadlineTask = parseDeadlineTaskCreationCommand(command);
-                handleAddingTask(deadlineTask);
-            } else if (command.startsWith(Command.EVENT.toString())) {
-                Task eventTask = parseEventTaskCreationCommand(command);
-                handleAddingTask(eventTask);
-            } else {
-                throw new UnknownInput();
-            }
-        } catch (MarkTaskInputException | UnmarkTaskInputException | InvalidTask | TaskDoesNotExist | UnknownInput | DeleteTaskInputException e) {
-            System.out.println(e.getMessage());
-        }
+        this.commandParser.parseAndHandle(command);
     }
 
     public void deleteTask(int taskNumber) throws TaskDoesNotExist {
         this.TASKS_LIST.deleteTask(taskNumber);
-    }
-
-    public int parseDeleteTaskCommand(String command) throws DeleteTaskInputException {
-        String[] commandArray = command.split(" ");
-        if (commandArray.length != 2) {
-            throw new DeleteTaskInputException();
-        }
-
-        try {
-            return Integer.parseInt(commandArray[1]);
-        } catch (NumberFormatException e) {
-            throw new DeleteTaskInputException();
-        }
-    }
-    public int parseMarkTaskCommand(String command) throws MarkTaskInputException {
-        String[] commandArray = command.split(" ");
-        if (commandArray.length != 2) {
-            throw new MarkTaskInputException();
-        }
-
-        try {
-            return Integer.parseInt(commandArray[1]);
-        } catch (NumberFormatException e) {
-            throw new MarkTaskInputException();
-        }
-    }
-
-    public int parseUnmarkTaskCommand(String command) throws UnmarkTaskInputException {
-        String[] commandArray = command.split(" ");
-        if (commandArray.length != 2) {
-            throw new UnmarkTaskInputException();
-        }
-
-        try {
-            return Integer.parseInt(commandArray[1]);
-        } catch (NumberFormatException e) {
-            throw new UnmarkTaskInputException();
-        }
-    }
-
-    public Task parseToDoTaskCreationCommand(String command) throws InvalidTask {
-        String[] commandArray = command.split(" ");
-        String[] taskDescriptionArr = Arrays.copyOfRange(commandArray, 1, commandArray.length);
-        String taskDescription = String.join(" ", taskDescriptionArr);
-        if (taskDescription.isEmpty()) {
-            throw new InvalidTask("Missing description for ToDo task");
-        }
-        return new ToDo(taskDescription);
-    }
-
-    public Task parseDeadlineTaskCreationCommand(String command) throws InvalidTask {
-        String[] commandArray = command.split(" ");
-
-        int byIndex = Arrays.asList(commandArray).indexOf("/by");
-        if (byIndex == -1) {
-            throw new InvalidTask("Missing `by` for Deadline task");
-        }
-
-        String[] taskDescriptionArr = Arrays.copyOfRange(commandArray, 1, byIndex);
-        String taskDescription = String.join(" ", taskDescriptionArr);
-        if (taskDescription.isEmpty()) {
-            throw new InvalidTask("Missing description for Deadline task");
-        }
-
-        String[] byDateArr = Arrays.copyOfRange(commandArray, byIndex + 1, commandArray.length);
-        String byDate = String.join(" ", byDateArr);
-        if (byDate.isEmpty()) {
-            throw new InvalidTask("Missing by date for Deadline task");
-        }
-
-        Task deadlineTask;
-
-        try {
-            deadlineTask = new Deadline(taskDescription, LocalDateTimeHandler.parseLocalDateTime(byDate));
-        } catch (DateTimeParseException e) {
-            throw new InvalidTask("Invalid deadline date");
-        }
-        return deadlineTask;
-    }
-
-    public Task parseEventTaskCreationCommand(String command) throws InvalidTask {
-        String[] commandArray = command.split(" ");
-
-        int fromIndex = Arrays.asList(commandArray).indexOf("/from");
-        if (fromIndex == -1) {
-            throw new InvalidTask("Missing `from` for Event task");
-        }
-
-        String[] taskDescriptionArr = Arrays.copyOfRange(commandArray, 1, fromIndex);
-        String taskDescription = String.join(" ", taskDescriptionArr);
-        if (taskDescription.isEmpty()) {
-            throw new InvalidTask("Missing description for Event task");
-        }
-
-
-        int toIndex = Arrays.asList(commandArray).indexOf("/to");
-        if (toIndex == -1) {
-            throw new InvalidTask("Missing `to` for Event task");
-        }
-
-        String[] fromDateArr = Arrays.copyOfRange(commandArray, fromIndex + 1, toIndex);
-        String fromDate = String.join(" ", fromDateArr);
-        if (fromDate.isEmpty()) {
-            throw new InvalidTask("Missing from date for Event task");
-        }
-
-        String[] toDateArr = Arrays.copyOfRange(commandArray, toIndex + 1, commandArray.length);
-        String toDate = String.join(" ", toDateArr);
-        if (toDate.isEmpty()) {
-            throw new InvalidTask("Missing to date for Event task");
-        }
-
-        Task eventTask;
-        try {
-            eventTask = new Event(taskDescription, LocalDateTimeHandler.parseLocalDateTime(fromDate),
-                    LocalDateTimeHandler.parseLocalDateTime(toDate));
-        } catch (DateTimeParseException e) {
-            throw new InvalidTask("Invalid event date(s)");
-        }
-        return eventTask;
     }
 
     public void handleAddingTask(Task task) {
